requirements.txt
Flask==1.1.2
opencv-python==4.5.2.54
Pillow==8.1.0
face-recognition==1.3.0
dlib==19.22.0